# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ipoh Drum Laravel API",
        "description": "This is a swagger-generated API documentation for the project Ipoh Drum. (Only supports OpenAPI Annotations for now.)",
        "contact": {
            "email": "henry_lcz97@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "The URL that IpohDrum Laravel API is running on."
        }
    ],
    "paths": {
...
        "/api/ticket": {
            "get": {
                "tags": [
                    "TicketControllerService"
                ],
                "summary": "Get list of tickets",
                "description": "Returns list of tickets",
                "operationId": "getTickets",
                "parameters": [
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "number of pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of tickets"
                    },
                    "default": {
                        "description": "Unable to retrieve list of tickets"
                    }
                }
            },
            "post": {
                "tags": [
                    "TicketControllerService"
                ],
                "summary": "Creates a ticket.",
                "operationId": "createTicket",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Ticketname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "storeid",
                        "in": "query",
                        "description": "Store ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "promotionid",
                        "in": "query",
                        "description": "Promotion ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Code",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sku",
                        "in": "query",
                        "description": "Sku",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Product Description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "imgpath",
                        "in": "query",
                        "description": "Image Path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Product Selling Price",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "qty",
                        "in": "query",
                        "description": "Stock Qty",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stockthreshold",
                        "in": "query",
                        "description": "Stock Threshold",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "enddate",
                        "in": "query",
                        "description": "Valid end date",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "onsale",
                        "in": "query",
                        "description": "On Sale",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket has been created successfully."
                    },
                    "default": {
                        "description": "Unable to create the ticket."
                    }
                }
            }
        },
        "/api/filter/ticket": {
            "get": {
                "tags": [
                    "TicketControllerService"
                ],
                "summary": "Filter list of tickets",
                "description": "Returns list of filtered tickets",
                "operationId": "filterTickets",
                "parameters": [
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "number of pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromdate",
                        "in": "query",
                        "description": "From Date for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "todate",
                        "in": "query",
                        "description": "To date for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "onsale",
                        "in": "query",
                        "description": "On sale for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status for filter",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of filtered tickets"
                    },
                    "default": {
                        "description": "Unable to retrieve list of tickets"
                    }
                }
            }
        },
...
    },
...
}